Improve File Menu + :recycle:
Remove import and export dialogs and refine file menu and util
Implement native file browser and refactor Import Export functionality
Update menu ordering and improve submenu top margin settings
