+++ package.json
@@ -36,6 +36,7 @@
         "@reduxjs/toolkit": "^1.5.0",
         "axios": "^0.24.0",
         "bfj": "^7.0.2",
+        "browser-fs-access": "^0.24.0",
         "browserslist": "^4.18.1",
         "camelcase": "^6.2.1",
         "canvas": "^2.8.0",
@@ -113,6 +114,7 @@
         "@types/react-transition-group": "^4.4.4",
         "@types/sat": "^0.0.31",
         "@types/styled-components": "^5.1.12",
+        "@types/wicg-file-system-access": "^2020.9.4",
         "@typescript-eslint/eslint-plugin": "^5.9.1",
         "@typescript-eslint/parser": "^5.4.0",
         "autoprefixer": "^10.0.2",
