+++ src/drawing/io/helpers.ts
@@ -1,0 +1,11 @@
+export const readFileAsUint8Array = async (file: File): Promise<Uint8Array> =>
+    new Promise((resolve, reject) => {
+        const fileReader = new FileReader()
+        fileReader.onloadend = () => {
+            resolve(new Uint8Array(fileReader.result as ArrayBuffer))
+        }
+        fileReader.onerror = (err) => {
+            reject(err)
+        }
+        fileReader.readAsArrayBuffer(file)
+    })
