+++ src/language/en.json
@@ -14,10 +14,9 @@
     "Menu.General.About": "About...",
     "Menu.General.File.New": "New",
     "Menu.General.File.Open": "Open...",
-    "Menu.General.File.Save": "Save",
-    "Menu.General.File.SaveAs": "Save As...",
-    "Menu.General.File.Import": "Import File",
-    "Menu.General.File.Export": "Export File",
+    "Menu.General.File.Save": "Save...",
+    "Menu.General.File.ImportPdf": "Import PDF...",
+    "Menu.General.File.ExportPdf": "Export as PDF...",
     "Menu.General.Edit.Undo": "Undo",
     "Menu.General.Edit.Redo": "Redo",
     "Menu.Page.Style": "Paper Style",
@@ -83,11 +82,10 @@
     "SessionMenu.CreateOnline.JoinButton": "Join Session",
     "SessionMenu.CreateOnline.TextFieldLabel.SessionId": "Insert Session ID",
     "SessionMenu.CreateOnline.InvalidSession": "Looks like the session you're trying to join does not exist 🤖",
-    "SessionMenu.InitialSelection.Title": "Create Session",
-    "SessionMenu.InitialSelection.CreateOffline": "Create offline session",
-    "SessionMenu.InitialSelection.OnlineSessionMenu": "Online session menu",
-    "SessionMenu.InitialSelection.ImportFile": "Import file",
-    "SessionMenu.InitialSelection.Continue": "Continue previous session",
+    "SessionMenu.InitialSelection.Title": "Welcome to BoardSite.io",
+    "SessionMenu.InitialSelection.CreateOffline": "Create Offline Session",
+    "SessionMenu.InitialSelection.OnlineSessionMenu": "Create Online Session",
+    "SessionMenu.InitialSelection.Continue": "Continue Previous Session",
     "SessionMenu.JoinOnly.Title": "Collaborative Session 👋🏻",
     "SessionMenu.JoinOnly.JoinButton": "Join Session",
     "SessionMenu.JoinOnly.UnableToJoin": "Unable to join session!",
